+++ mayan/apps/mirroring/tests/literals.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, unicode_literals
+
+TEST_NODE_EXPRESSION = 'level_1'
+TEST_NODE_EXPRESSION_MULTILINE = 'first\r\nsecond\r\nthird'